# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for Objective Zero"
  version: "1.0.0"
  title: "Objective Zero API"
  contact:
    email: "you@your-company.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "107.170.82.21:3000"
basePath: "/mobile"
...
schemes:
- "http"
paths:
...
  /assignUserRole:
    post:
      summary: "Change the Role of the User"
      description: "Assign the user role. Role must be Ambassador, Veteran or user."
      operationId: "assignUserRole"
      produces:
      - "application/json"
      parameters:
        - name: access_token
          in: header
          description: Access Token
          required: true
          type: string
          format: string
        - in: body
          name: Assign User Role
          description: Assign User a role or update.
          schema:
            type: object
            required:
              - user_role
            properties:
              user_role:
                type: string 
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "user_detail"
...
  /ambassadorData:
    post:
      summary: "Get List of All Ambassador"
      description: "User can get the lost of all ambassador available in our record"
      operationId: "ambassadorData"
      produces:
      - "application/json"
      parameters:
        - name: access_token
          in: header
          description: Access Token
          required: true
          type: string
          format: string            
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "add_ambassador"
...